+++ src/Endpoint/Configuration/DbContextManagerBehavior.cs
