+++ src/Endpoint/Configuration/PersistenceConfiguration.cs
