+++ src/Endpoint/Configuration/RecoveryConfiguration.cs
