+++ src/Endpoint/Configuration/SerializationConfiguration.cs
