+++ src/Endpoint/Endpoint.csproj
