+++ src/Endpoint/EndpointConfiguration.cs
