+++ src/Endpoint/Handlers/PlaceOrderCommandHandler.cs
