+++ src/Endpoint/NServiceBus.Hosting.Azure.HostProcess.exe.config
