+++ src/Endpoint/Properties/AssemblyInfo.cs
