+++ src/UnitTests/UnitTests.csproj
@@ -81,7 +81,7 @@
       <Project>{11B8C825-077F-4726-A55C-842277850A32}</Project>
       <Name>Domain</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Endpoint2\Endpoint.csproj">
+    <ProjectReference Include="..\Endpoint\Endpoint.csproj">
       <Project>{0f372666-a717-4cb9-b459-4509cec47c55}</Project>
       <Name>Endpoint</Name>
     </ProjectReference>
